backend/requirements.txt
langgraph==1.0.10
langchain==1.2.10
openai==2.26.0
fastapi==0.135.1
uvicorn==0.41.0
pydantic==2.12.5
python-dotenv==1.2.2
requests==2.32.5
beautifulsoup4==4.14.3
httpx==0.28.1
tiktoken==0.12.0
langchain-openai==1.1.10
uvicorn[standard]==0.41.0